# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cadastro de Livros API",
        "description": "API REST para gerenciamento de livros, autores e assuntos.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/autores": {
            "get": {
                "tags": [
                    "Autores"
                ],
                "summary": "Listar autores (paginado)",
                "operationId": "eb27430fea956ab5ff5aaca0e1355868",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de autores",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Autor"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Autores"
                ],
                "summary": "Criar autor",
                "operationId": "79b1e52e74880ae8637d083d4705c17d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AutorInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Autor criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Autor"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
...
        "/autores/{id}": {
            "get": {
                "tags": [
                    "Autores"
                ],
                "summary": "Buscar autor",
                "operationId": "dd01080f500d7bbcb4d41eba034ebbff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Autor encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Autor"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Autores"
                ],
                "summary": "Atualizar autor",
                "operationId": "9dc6b2621eb50caac0dc2bcd21888aa8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AutorInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Autor atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Autor"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não encontrado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Autores"
                ],
                "summary": "Remover autor",
                "operationId": "fe487831afee2422c6afc92243a36d1a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Removido"
                    },
                    "404": {
                        "description": "Não encontrado"
                    },
                    "422": {
                        "description": "Autor vinculado a livros"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Autor": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nome": {
                        "type": "string",
                        "example": "Machado de Assis"
                    }
                },
                "type": "object"
            },
            "AutorInput": {
                "required": [
                    "nome"
                ],
                "properties": {
                    "nome": {
                        "type": "string",
                        "example": "Machado de Assis",
                        "maxLength": 40
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}